[backup] cli: refactor: move ReadRecordBytes to parent mod
