+++ storage/backup/backup-cli/src/backup.rs
@@ -1,15 +1,13 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{adapter::Adapter, FileHandle};
-use anyhow::{bail, Result};
-use async_trait::async_trait;
-use bytes::{Bytes, BytesMut};
+use crate::{adapter::Adapter, FileHandle, ReadRecordBytes};
+use anyhow::Result;
+use bytes::Bytes;
 use futures::{stream, stream::TryStreamExt};
 use libra_crypto::HashValue;
 use libra_types::{account_state_blob::AccountStateBlob, transaction::Version};
-use std::convert::TryInto;
-use tokio::io::AsyncReadExt;
+use tokio::io::{AsyncRead, AsyncReadExt};
 use tokio_util::compat::FuturesAsyncReadCompatExt;
 
 pub struct BackupServiceClient {
@@ -25,7 +23,7 @@ impl BackupServiceClient {
         }
     }
 
-    async fn get(&self, path: &str) -> Result<impl AsyncReadExt> {
+    async fn get(&self, path: &str) -> Result<impl AsyncRead> {
         Ok(self
             .client
             .get(&format!("http://localhost:{}/{}", self.port, path))
@@ -51,65 +49,13 @@ impl BackupServiceClient {
         &self,
         key: HashValue,
         version: Version,
-    ) -> Result<impl AsyncReadExt> {
+    ) -> Result<impl AsyncRead> {
         self.get(&format!("state_range_proof/{}/{:x}", version, key))
             .await
     }
 
-    async fn get_state_snapshot(&self, version: Version) -> Result<impl AsyncReadExt> {
+    async fn get_state_snapshot(&self, version: Version) -> Result<impl AsyncRead> {
         self.get(&format!("state_snapshot/{}", version)).await
-    }
-}
-
-#[async_trait]
-trait ReadRecordBytes {
-    async fn read_full_buf_or_none(&mut self, buf: &mut BytesMut) -> Result<()>;
-    async fn read_record_bytes(&mut self) -> Result<Option<Bytes>>;
-}
-
-#[async_trait]
-impl<R: AsyncReadExt + Send + Unpin> ReadRecordBytes for R {
-    async fn read_full_buf_or_none(&mut self, buf: &mut BytesMut) -> Result<()> {
-        assert_eq!(buf.len(), 0);
-        let n_expected = buf.capacity();
-
-        loop {
-            let n_read = self.read_buf(buf).await?;
-            let n_read_total = buf.len();
-            if n_read_total == n_expected {
-                return Ok(());
-            }
-            if n_read == 0 {
-                if n_read_total == 0 {
-                    return Ok(());
-                } else {
-                    bail!(
-                        "Hit EOF before filling the whole buffer, read {}, expected {}",
-                        n_read_total,
-                        n_expected
-                    );
-                }
-            }
-        }
-    }
-
-    async fn read_record_bytes(&mut self) -> Result<Option<Bytes>> {
-        // read record size
-        let mut size_buf = BytesMut::with_capacity(4);
-        self.read_full_buf_or_none(&mut size_buf).await?;
-        if size_buf.is_empty() {
-            return Ok(None);
-        }
-
-        // read record
-        let record_size = u32::from_be_bytes(size_buf.as_ref().try_into()?) as usize;
-        let mut record_buf = BytesMut::with_capacity(record_size);
-        self.read_full_buf_or_none(&mut record_buf).await?;
-        if record_buf.is_empty() {
-            bail!("Hit EOF when reading record.")
-        }
-
-        Ok(Some(record_buf.freeze()))
     }
 }
